requirements.txt
requests
beautifulsoup4
newspaper3k
groq
python-dotenv